Tehtävät 4.15-4.21

+++ osa4/controllers/blogs.js
@@ -1,5 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 /**
  * Hakee tietokannasta ja palauttaa kaikki blogiartikkelit json-muodossa.
@@ -7,7 +9,10 @@ const Blog = require('../models/blog')
 blogsRouter.get('/', async (request, response) => {
     try {
         const blogs = await Blog.find({})
-        response.json(blogs)
+            .populate('user', { username: 1, name: 1 })
+
+        response.json(blogs.map(Blog.format))
+
     } catch (exception) {
         console.log(exception)
         response.status(500).json({ error: 'something went wrong' })
@@ -20,7 +25,10 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.get('/:id', async (request, response) => {
     try {
         const blog = await Blog.findById(request.params.id)
-        response.json(blog)
+            .populate('user', { username: 1, name: 1 })
+
+        response.json(Blog.format(blog))
+
     } catch (exception) {
         console.log(exception)
         response.status(500).json({ error: 'something went wrong' })
@@ -31,25 +39,46 @@ blogsRouter.get('/:id', async (request, response) => {
  * Tallettaa uuden blogiartikkelin tietokantaan.
  */
 blogsRouter.post('/', async (request, response) => {
-    try {
-        const blog = new Blog({
-            title: request.body.title,
-            author: request.body.author,
-            url: request.body.url,
-            likes: request.body.likes === undefined ? 0 : request.body.likes
-        })
+    const body = request.body
 
-        // ilman otsikkoa tai urlia blogia ei talleteta
-        if (blog.title === undefined || blog.url === undefined) {
-            return response.status(400).json({ error: 'blog should have title and url' })
+    // ilman otsikkoa tai urlia ei tehdä mitään
+    if (body.title === undefined || body.url === undefined) {
+        return response.status(400).json({ error: 'blog should have title and url' })
+    }
+
+    try {
+        // tarkistetaan tokenin oikeellisuus
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        // haetaan käyttäjä tokenin avulla
+        const user = await User.findById(decodedToken.id)
+
+        if (!request.token || !decodedToken.id || !user) {
+            return response.status(401).json({ error: 'token missing or invalid' })
         }
 
+        const blog = new Blog({
+            user: user.id,
+            likes: body.likes === undefined ? 0 : request.body.likes,
+            author: body.author,
+            title: body.title,
+            url: body.url
+        })
+
         const savedBlog = await blog.save()
+
+        // talletetaan blogi myös käyttäjän tietoihin
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
+
         response.status(201).json(savedBlog)
 
     } catch (exception) {
-        console.log(exception)
-        response.status(500).json({ error: 'something went wrong' })
+        if (exception.name === 'JsonWebTokenError') {
+            response.status(401).json({ error: exception.message })
+        } else {
+            console.log(exception)
+            response.status(500).json({ error: 'something went wrong' })
+        }
     }
 })
 
@@ -58,11 +87,39 @@ blogsRouter.post('/', async (request, response) => {
  */
 blogsRouter.delete('/:id', async (request, response) => {
     try {
+        // tarkistetaan tokenin oikeellisuus
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        const removerId = decodedToken.id
+
+        if (!request.token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+        // tarkistetaan että poistaja on sama kuin blogin lisääjä
+        const blogToRemove = await Blog.findById(request.params.id)
+
+        const userId = blogToRemove.user
+        if (removerId.toString() !== userId.toString()) {
+            return response.status(401).json({ error: 'not authorized to delete this blog' })
+        }
+
+        // poistetaan blogi myös käyttäjän tiedoista
+        const user = await User.findById(userId)
+        user.blogs = user.blogs.filter(blog => {
+            return blog.toString() !== request.params.id
+        })
+        await user.save()
+
         await Blog.findByIdAndDelete(request.params.id)
         response.status(204).end()
+
     } catch (exception) {
-        console.log(exception)
-        response.status(500).json({ error: 'something went wrong' })
+        if (exception.name === 'JsonWebTokenError') {
+            response.status(401).json({ error: exception.message })
+        } else {
+            console.log(exception)
+            response.status(500).json({ error: 'something went wrong' })
+        }
     }
 })
